Introduce Han2005 model
