+++ pooltool/physics/engine.py
@@ -4,10 +4,6 @@ import attrs
 
 from pooltool.events.datatypes import AgentType, Event, EventType
 from pooltool.physics.resolve import Resolver
-from pooltool.physics.resolve.ball_cushion import (
-    resolve_circular_ball_cushion,
-    resolve_linear_ball_cushion,
-)
 from pooltool.physics.resolve.ball_pocket import resolve_ball_pocket
 from pooltool.physics.resolve.stick_ball import resolve_stick_ball
 from pooltool.system.datatypes import System
@@ -58,12 +54,14 @@ class PhysicsEngine:
         elif event.event_type == EventType.BALL_LINEAR_CUSHION:
             ball = shot.balls[ids[0]]
             cushion = shot.table.cushion_segments.linear[ids[1]]
-            _ = resolve_linear_ball_cushion(ball, cushion, inplace=True)
+            _ = self.resolver.ball_linear_cushion.resolve(ball, cushion, inplace=True)
             ball.state.t = event.time
         elif event.event_type == EventType.BALL_CIRCULAR_CUSHION:
             ball = shot.balls[ids[0]]
             cushion_jaw = shot.table.cushion_segments.circular[ids[1]]
-            _ = resolve_circular_ball_cushion(ball, cushion_jaw, inplace=True)
+            _ = self.resolver.ball_circular_cushion.resolve(
+                ball, cushion_jaw, inplace=True
+            )
             ball.state.t = event.time
         elif event.event_type == EventType.BALL_POCKET:
             ball = shot.balls[ids[0]]
