+++ pooltool/physics/resolve/ball_cushion/han_2005/README.md
@@ -1,0 +1,1 @@
+https://ekiefl.github.io/2020/04/24/pooltool-theory/#3-han-2005
